+++ cmsplugin_blocks/factories/card.py
@@ -3,6 +3,7 @@ import random
 import factory
 
 from cmsplugin_blocks.choices_helpers import get_card_default_template
+from cmsplugin_blocks.factories.utils import create_image_file
 from cmsplugin_blocks.models import Card
 
 
@@ -12,7 +13,6 @@ class CardFactory(factory.django.DjangoModelFactory):
     """
     template = get_card_default_template()
     content = factory.Faker("text", max_nb_chars=42)
-    image = factory.django.FileField(filename="foo.jpg")
 
     class Meta:
         model = Card
@@ -23,3 +23,14 @@ class CardFactory(factory.django.DjangoModelFactory):
         Select a random alignment
         """
         return random.choice([item[0] for item in Card.ALIGNMENT_CHOICES])
+
+    @factory.lazy_attribute
+    def image(self):
+        """
+        Fill file field with generated image on the fly by PIL.
+
+        Returns:
+            django.core.files.File: File object.
+        """
+
+        return create_image_file()