+++ tests/100_element_card.py
@@ -3,11 +3,15 @@ import re
 
 import pytest
 
+from cms.api import create_page, add_plugin
+from cms.utils.urlutils import admin_reverse
+
 from tests.utils import FixturesTestCaseMixin, CMSPluginTestCase
 
 from cmsplugin_blocks.choices_helpers import get_card_default_template
 from cmsplugin_blocks.cms_plugins import CardPlugin
 from cmsplugin_blocks.factories.card import CardFactory
+from cmsplugin_blocks.factories.user import UserFactory
 
 
 def test_factory(db):
@@ -53,14 +57,14 @@ class CardCMSPluginsTestCase(FixturesTestCaseMixin, CMSPluginTestCase):
         Every parts should be present when rendered
         """
         # Create random values for parameters with a factory
-        fabricated = CardFactory(content="<p>Lorem ipsum dolore</p>")
+        card = CardFactory(content="<p>Lorem ipsum dolore</p>")
 
         placeholder, model_instance, context, html = self.create_basic_render(
             CardPlugin,
-            template=fabricated.template,
-            alignment=fabricated.alignment,
-            image=fabricated.image,
-            content=fabricated.content,
+            template=card.template,
+            alignment=card.alignment,
+            image=card.image,
+            content=card.content,
         )
 
         print()
@@ -70,7 +74,7 @@ class CardCMSPluginsTestCase(FixturesTestCaseMixin, CMSPluginTestCase):
         pattern = (
             r'<div class="card card--{}">'
         ).format(
-            fabricated.alignment
+            card.alignment
         )
         self.assertIsNotNone(re.search(pattern, html))
 
@@ -88,7 +92,7 @@ class CardCMSPluginsTestCase(FixturesTestCaseMixin, CMSPluginTestCase):
                 raise AssertionError("There is some 'sorl.thumbnail' error(s)")
 
         expected_content = """<div class="card__content">{}</div>""".format(
-            fabricated.content
+            card.content
         )
         self.assertInHTML(expected_content, html)
 
@@ -97,16 +101,16 @@ class CardCMSPluginsTestCase(FixturesTestCaseMixin, CMSPluginTestCase):
         When card has no image, image part should not be rendered
         """
         # Create random values for parameters with a factory
-        fabricated = CardFactory(
+        card = CardFactory(
             image=None
         )
 
         placeholder, model_instance, context, html = self.create_basic_render(
             CardPlugin,
-            template=fabricated.template,
-            alignment=fabricated.alignment,
-            image=fabricated.image,
-            content=fabricated.content,
+            template=card.template,
+            alignment=card.alignment,
+            image=card.image,
+            content=card.content,
         )
 
         print()
@@ -121,7 +125,7 @@ class CardCMSPluginsTestCase(FixturesTestCaseMixin, CMSPluginTestCase):
         self.assertIsNone(re.search(pattern, html))
 
         expected_content = """<div class="card__content">{}</div>""".format(
-            fabricated.content
+            card.content
         )
         self.assertInHTML(expected_content, html)
 
@@ -130,16 +134,16 @@ class CardCMSPluginsTestCase(FixturesTestCaseMixin, CMSPluginTestCase):
         When card has no content, content part should not be rendered
         """
         # Create random values for parameters with a factory
-        fabricated = CardFactory(
+        card = CardFactory(
             content=""
         )
 
         placeholder, model_instance, context, html = self.create_basic_render(
             CardPlugin,
-            template=fabricated.template,
-            alignment=fabricated.alignment,
-            image=fabricated.image,
-            content=fabricated.content,
+            template=card.template,
+            alignment=card.alignment,
+            image=card.image,
+            content=card.content,
         )
 
         print()
@@ -149,3 +153,146 @@ class CardCMSPluginsTestCase(FixturesTestCaseMixin, CMSPluginTestCase):
             r'<div class="card__content">'
         )
         self.assertIsNone(re.search(pattern, html))
+
+    def test_card_plugin_form_add(self):
+        """
+        Plugin creation form should return a success status code and every
+        expected field should be present in HTML.
+        """
+        # Connect a dummy admin
+        staff = UserFactory(is_staff=True, is_superuser=True)
+        self.client.login(username=staff.username, password="password")
+
+        # Create dummy page
+        page = create_page(
+            language="en",
+            title="Dummy",
+            slug="dummy",
+            template="pages/default.html",
+        )
+
+        # Get placeholder
+        placeholder = page.placeholders.get(slot="content")
+
+        # Get the edition plugin form url and open it
+        url = admin_reverse('cms_page_add_plugin')
+        response = self.client.get(url, {
+            'plugin_type': 'CardPlugin',
+            'placeholder_id': placeholder.pk,
+            'target_language': 'en',
+            'plugin_language': 'en',
+        })
+
+        html = response.content.decode("utf-8")
+
+        # Expected http success status
+        self.assertEqual(response.status_code, 200)
+
+        # Check all expected fields are present
+        self.assertIsNotNone(
+            re.search(
+                (
+                    r'<select.*id="id_alignment".*>'
+                ),
+                html
+            )
+        )
+        self.assertIsNotNone(
+            re.search(
+                (
+                    r'<select.*id="id_template".*>'
+                ),
+                html
+            )
+        )
+        self.assertIsNotNone(
+            re.search(
+                (
+                    r'<input.*type="file.*id="id_image".*>'
+                ),
+                html
+            )
+        )
+        self.assertIsNotNone(
+            re.search(
+                (
+                    r'<textarea.*id="id_content".*>'
+                ),
+                html
+            )
+        )
+
+    def test_card_plugin_form_edit(self):
+        """
+        Plugin edition form should return a success status code and every
+        expected field should be present in HTML.
+        """
+        # Create random values for parameters with a factory
+        card = CardFactory(content="<p>Lorem ipsum dolore</p>")
+
+        # Connect a dummy admin
+        staff = UserFactory(is_staff=True, is_superuser=True)
+        self.client.login(username=staff.username, password="password")
+
+        # Create dummy page
+        page = create_page(
+            language="en",
+            title="Dummy",
+            slug="dummy",
+            template="pages/default.html",
+        )
+
+        # Add card plugin to placeholder
+        placeholder = page.placeholders.get(slot="content")
+        model_instance = add_plugin(
+            placeholder,
+            CardPlugin,
+            "en",
+            template=card.template,
+            alignment=card.alignment,
+            image=card.image,
+            content=card.content,
+        )
+
+        # Get the edition plugin form url and open it
+        url = admin_reverse('cms_page_edit_plugin', args=[model_instance.id])
+        response = self.client.get(url)
+
+        html = response.content.decode("utf-8")
+
+        # Expected http success status
+        self.assertEqual(response.status_code, 200)
+
+        # Check all expected fields are present
+        self.assertIsNotNone(
+            re.search(
+                (
+                    r'<select.*id="id_alignment".*>'
+                ),
+                html
+            )
+        )
+        self.assertIsNotNone(
+            re.search(
+                (
+                    r'<select.*id="id_template".*>'
+                ),
+                html
+            )
+        )
+        self.assertIsNotNone(
+            re.search(
+                (
+                    r'<input.*type="file.*id="id_image".*>'
+                ),
+                html
+            )
+        )
+        self.assertIsNotNone(
+            re.search(
+                (
+                    r'<textarea.*id="id_content".*>'
+                ),
+                html
+            )
+        )