+++ tests/102_element_slider.py
@@ -236,7 +236,7 @@ class SliderCMSPluginsTestCase(FixturesTestCaseMixin, CMSPluginTestCase):
 
         assert items == [('1', 1), ('2', 2), ('3', 3), ('4', 4)]
 
-    def test_slider_plugin_form_add(self):
+    def test_slider_plugin_form_view_add(self):
         """
         Plugin creation form should return a success status code and every
         expected field should be present in HTML.
@@ -336,7 +336,7 @@ class SliderCMSPluginsTestCase(FixturesTestCaseMixin, CMSPluginTestCase):
             )
         )
 
-    def test_slider_plugin_form_edit(self):
+    def test_slider_plugin_form_view_edit(self):
         """
         Plugin edition form should return a success status code and every
         expected field should be present in HTML.