+++ cmsplugin_blocks/admin.py
@@ -5,7 +5,7 @@ Admin interface definitions
 import logging
 
 from django.contrib import admin
-from django.utils.translation import ugettext_lazy as _
+from django.utils.translation import gettext_lazy as _
 
 from sorl.thumbnail.helpers import ThumbnailError
 from sorl.thumbnail import get_thumbnail