+++ cmsplugin_blocks/forms/album.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 from django import forms
-from django.utils.translation import ugettext_lazy as _
+from django.utils.translation import gettext_lazy as _
 from django.forms.widgets import FileInput, NumberInput
 
 from cmsplugin_blocks.models.album import Album, AlbumItem