+++ cmsplugin_blocks/models/hero.py
@@ -5,7 +5,7 @@ from django.db import models
 from django.utils.encoding import force_text
 from django.utils.html import strip_tags
 from django.utils.text import Truncator
-from django.utils.translation import ugettext_lazy as _
+from django.utils.translation import gettext_lazy as _
 
 from cms.models.pluginmodel import CMSPlugin