+++ cmsplugin_blocks/plugins/slider.py
@@ -2,7 +2,7 @@
 """
 Slider CMS Plugin interface definitions
 """
-from django.utils.translation import ugettext_lazy as _
+from django.utils.translation import gettext_lazy as _
 
 from cms.plugin_base import CMSPluginBase