+++ HISTORY.rst
@@ -8,6 +8,7 @@ Version 0.6.0 - Unreleased
 * Added test coverage and factories for every plugins;
 * Modified every plugin templates to be cleaner and flawless;
 * Added ``title`` attribute to SlideItem model;
+* Use **sorl-thumbnail** on images in every plugin templates;
 
 Version 0.5.2 - 2019/05/18
 --------------------------