+++ HISTORY.rst
@@ -5,12 +5,12 @@ Changelog
 Version 0.6.0 - Unreleased
 --------------------------
 
-* Added test coverage and factories for every plugins;
+* Added test coverage for every plugins;
 * Added tox configuration;
 * Modified every plugin templates to be cleaner and flawless;
 * Added ``title`` attribute to SlideItem model;
 * Use **sorl-thumbnail** on images in every plugin templates;
-* Use FileField instead of ImageField for image fields in every plugin;
+* Use FileField instead of ImageField for image fields in every plugin so we can use SVG;
 
 Version 0.5.2 - 2019/05/18
 --------------------------